Update Objective-C and Swift to match Xcode.gitignore
The Objective-C and Swift gitignore files are missing `*.xccheckout`
and `*.xcscmblueprint`, but these are still included in the Xcode
gitignore. This commit synchronises the three files.

The files themselves contain information about the user's local
source control status and shouldn't be included in most Xcode
repositories.

`*.xcuserstate` appears to always be contained in `xcuserdata/`
which is already ignored.