Ignore *.xyd outline files generated by xy LaTeX package (#3088)
The xy package generates *.xyd files whenever the commands

    \MakeOutlines
    \OnlyOutlines
    \ShowOutlines
    \NoOutlines

are present in the LaTeX source. These automatically-generated files
contain the dimensions of figures typeset with xy and are recreated as
needed. This is documented on pp. 15f. of the XY-pic Reference Manual
(1999/02/16).

As automatically-generated, temporary files, they should be ignored.